+++ plugin-boilerplate/{{ cookiecutter.project_slug }}/plugins/{{ cookiecutter.project_module }}/backend/src/{{ cookiecutter.project_module }}/plugins.py
@@ -1,19 +1,19 @@
 import logging
-
+from baserow.core.registries import Plugin
 from django.urls import path, include
 
-from baserow.core.registries import Plugin
-
 from .api import urls as api_urls
-
 
 logger = logging.getLogger(__name__)
 
 
 class PluginNamePlugin(Plugin):
-    type = '{{ cookiecutter.project_slug }}'
+    type = "{{ cookiecutter.project_module }}"
 
     def get_api_urls(self):
         return [
-            path('{{ cookiecutter.project_slug }}/', include(api_urls, namespace=self.type)),
+            path(
+                "{{ cookiecutter.project_module }}/",
+                include(api_urls, namespace=self.type),
+            ),
         ]