+++ plugin-boilerplate/{{ cookiecutter.project_slug }}/settings/base.py
@@ -1,4 +1,0 @@
-from baserow.config.settings.base import *
-
-
-INSTALLED_APPS = INSTALLED_APPS + ['{{ cookiecutter.project_module }}']