+++ plugin-boilerplate/{{ cookiecutter.project_slug }}/settings/dev.py
@@ -1,5 +1,0 @@
-from .base import *
-
-
-DEBUG = True
-EMAIL_BACKEND = 'django.core.mail.backends.console.EmailBackend'